Оптимизация: выносит состояние из компонента Карта и добавляет функциональным компонентам memo
